+++ mod/src/main/java/basemod/patches/whatmod/RelicView.java
@@ -1,11 +1,12 @@
 package basemod.patches.whatmod;
 
+import basemod.ReflectionHacks;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.megacrit.cardcrawl.core.Settings;
+import com.megacrit.cardcrawl.helpers.Hitbox;
 import com.megacrit.cardcrawl.relics.AbstractRelic;
 import com.megacrit.cardcrawl.screens.SingleRelicViewPopup;
-
-import java.lang.reflect.Field;
 
 @SpirePatch(
     clz=SingleRelicViewPopup.class,
@@ -16,15 +17,10 @@ public class RelicView
     public static void Postfix(SingleRelicViewPopup __instance, SpriteBatch sb)
     {
         if (WhatMod.enabled) {
-            try {
-                Field relicField = SingleRelicViewPopup.class.getDeclaredField("relic");
-                relicField.setAccessible(true);
-                AbstractRelic relic = (AbstractRelic) relicField.get(__instance);
+			AbstractRelic relic = ReflectionHacks.getPrivate(__instance, SingleRelicViewPopup.class, "relic");
+            Hitbox nextHb = ReflectionHacks.getPrivate(__instance, SingleRelicViewPopup.class, "nextHb");
 
-                WhatMod.renderModTooltip(sb, relic.getClass());
-            } catch (IllegalAccessException | NoSuchFieldException e) {
-                e.printStackTrace();
-            }
-        }
+            WhatMod.renderModTooltipBottomLeft(sb, Settings.WIDTH / 2f + 340f * Settings.scale, nextHb.cY + 160f * Settings.scale, relic.getClass());
+		}
     }
 }