+++ CHANGELOG.md
@@ -1063,3 +1063,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 #### dev ####
 * Make basegame start relic upgrades replace the correct relic (erasels/kiooeht)
 * Make framebuffers nested (kiooeht)
+* Fix whatmod tooltip in the wrong position on SCV and SRV (kiooeht)